requirements.txt
asgiref==3.5.2
astroid==2.12.5
cffi==1.15.1
cryptography==37.0.4
dill==0.3.5.1
Django==4.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
flake8==5.0.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
pylint==2.15.0
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.4
wrapt==1.14.1
